mine/requirement.txt
numpy
opencv-python
Pillow
tensorflow
mss
matplotlib
scikit-learn
itertools